pyproject.toml
[tool.poetry]
name = "shade-proxy"
version = "0.1.1"
description = "Shade Proxy"
authors = ["Shade <shade@shade.inc>"]
license = "MIT"
packages = [{ include = "proxy" }]

[tool.poetry.scripts]
shade-proxy-tests = "tests.main:main"


[tool.poetry.dependencies]
python = "^3.8"
requests = "*"
google-cloud-storage = "*"
fastapi = "*"
uvicorn = "*"
orjson = "*"
shade-python-sdk = "*"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
